requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-cloudinary-storage==0.3.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
